Move stat tracking to a separate file
